Move Marshalling configuration out of EnvUtil
