+++ errai-cdi/errai-cdi-server/src/main/java/org/jboss/errai/cdi/server/events/EventDispatcher.java
@@ -41,7 +41,7 @@ import org.jboss.errai.bus.client.api.messaging.MessageCallback;
 import org.jboss.errai.bus.server.api.RpcContext;
 import org.jboss.errai.bus.server.util.LocalContext;
 import org.jboss.errai.common.client.protocols.MessageParts;
-import org.jboss.errai.config.rebind.EnvUtil;
+import org.jboss.errai.config.MarshallingConfiguration;
 import org.jboss.errai.enterprise.client.cdi.CDICommands;
 import org.jboss.errai.enterprise.client.cdi.CDIProtocol;
 import org.jboss.errai.enterprise.client.cdi.api.CDI;
@@ -96,7 +96,7 @@ public class EventDispatcher implements MessageCallback {
         final ClientObserverMetadata clientObserver = new ClientObserverMetadata(type, annotationTypes);
 
         if (!clientObservers.contains(clientObserver)) {
-          if (type == null || !EnvUtil.isPortableType(type)) {
+          if (type == null || !MarshallingConfiguration.isPortableType(type)) {
             log.warn("client tried to register a non-portable type: " + type);
             return;
           }
