+++ errai-common-apt/src/main/java/org/jboss/errai/common/apt/configuration/AptErraiConfiguration.java
@@ -16,7 +16,7 @@
 
 package org.jboss.errai.common.apt.configuration;
 
-import org.jboss.errai.common.apt.MetaClassFinder;
+import org.jboss.errai.config.MetaClassFinder;
 import org.jboss.errai.config.ErraiAppConfiguration;
 import org.jboss.errai.config.ErraiConfiguration;
 import org.jboss.errai.config.ErraiModulesConfiguration;
