+++ errai-common-apt/src/test/java/org/jboss/errai/common/apt/TestMetaClassFinder.java
@@ -17,6 +17,7 @@
 package org.jboss.errai.common.apt;
 
 import org.jboss.errai.codegen.meta.MetaClass;
+import org.jboss.errai.config.MetaClassFinder;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
