+++ errai-data-binding/src/main/java/org/jboss/errai/databinding/rebind/BindableProxyLoaderGenerator.java
@@ -33,7 +33,7 @@ import org.jboss.errai.codegen.meta.MetaClassFactory;
 import org.jboss.errai.codegen.meta.MetaParameterizedType;
 import org.jboss.errai.codegen.meta.MetaType;
 import org.jboss.errai.codegen.util.Stmt;
-import org.jboss.errai.common.apt.MetaClassFinder;
+import org.jboss.errai.config.MetaClassFinder;
 import org.jboss.errai.common.metadata.RebindUtils;
 import org.jboss.errai.config.ErraiAppPropertiesConfiguration;
 import org.jboss.errai.config.rebind.AbstractAsyncGenerator;
