+++ errai-marshalling/src/main/java/org/jboss/errai/marshalling/rebind/MarshallerGeneratorFactory.java
@@ -38,7 +38,6 @@ import org.jboss.errai.codegen.util.If;
 import org.jboss.errai.codegen.util.Stmt;
 import org.jboss.errai.common.rebind.NameUtil;
 import org.jboss.errai.common.rebind.UniqueNameGenerator;
-import org.jboss.errai.config.rebind.CommonConfigAttribs;
 import org.jboss.errai.marshalling.client.api.DeferredMarshallerCreationCallback;
 import org.jboss.errai.marshalling.client.api.GeneratedMarshaller;
 import org.jboss.errai.marshalling.client.api.Marshaller;
